refactor: Update layout manager in sequence widget
The code changes involve updating the layout manager in the sequence widget. The `SW_beat_frame.py` file has been modified to include a new method `adjust_layout_to_sequence_length()` which adjusts the layout of the beat frame based on the length of the sequence. This change improves the dynamic layout of the sequence widget.