Refactor splash screen initialization and display
This commit refactors the code related to the splash screen in the main.py and splash_screen.py files. The SplashScreenManager class in splash_screen.py is renamed to SplashScreen and inherits from QSplashScreen. The initialization of the splash screen is modified to use the QPixmap class and the show() method is renamed to show_message(). The commit also updates the usage of the splash screen in the main.py file by creating an instance of the SplashScreen class and calling the show_message() method. The finish() method is modified to call the finish() method of the base class. These changes improve the organization and readability of the code related to the splash screen.