Refactor prop classification and mapping
This commit refactors the prop classification and mapping by adding a new prop type "Triquetra" to the prop classifier and prop class mapping. It also updates the prop type tab in the settings dialog to include the "Triquetra" prop type. This change improves the organization and management of prop types in the codebase.